Restructured fs handler common code
-- Restructured the common code that supports each filesystem handler in
preparation for the introduction of multiple handlers based on the v7
filesystem layer (specifically, multiple Xenix filesystems).  As part of
this change, a new source file was created (v7fusecommon.c) containing code
that will be common across all v7-based handlers.  Note that, as of this
commit, only the v7fs handler has been updated to work with the new common
code.  Thus all the other handlers will fail to compile.

-- Restructured the parsing of configuration options to be cooperative between
the common code and filesystem-specific code.  The filesystem-specific code
is called first to process any fs unique options, or any whose behavior needs
to be different from that implemented by the common code.  After this, any
options that remain are passed to the common option processing code.

-- Extended the common code to support a set of 9 new configuration options
for the disk I/O layer: container, dsksize, dskoffset, cylinders, heads,
sectors, interleave, partnum and parttype.  A tenth option, dsklayout, was
implemented in the v7-specific code to handle parsing disk layout options.

-- Restructured the code surrounding filesystem initialization.  The “initfs”
option was renamed to “mkfs”, and parsing of this option was moved into common
code (although with the new cooperative option parsing scheme it can be over-
ridden by each handler).  The disk I/O layer is now initialized once, in
common code, before a call to fs-specific code to perform the actual filesystem
initialization, which is common for all the v7 variants.

-- Updated V7Tests.py to work with new v7fs command line options.  All v7fs
tests pass.

-- Fixed small typo in definition of Xenix 3 superblock.