Initial support for trsxenix disk layout
-- Added the ability to open and create disks that use the TRS-XENIX disk
layout.  The disk I/O layer can now read disk geometry and basic alternate
track information from the “.pvh” and “.bad” tables that are placed on the
disk by the TRS-XENIX diskutil tool.  Additionally, the disk I/O layer can
write new .pvh and .bad tables when initializing a new disk, based on
information supplied on the command line.  Note that, as of yet, there is
no support for honoring the entries in the bad track table.